Implement multithreaded version
We now have a render thread that just listens on a channel for
events to do: render or exit (for now, maybe resize separately
later?)

The event listener runs in a separate goroutine and whenever an
event occurs, it deals with it and pushes the appropriate
(render/exit) info on the channel.
This way we can also update the display when a weather query
finishes, without having to wait for it.

Lastly we have a goroutine that just waits until the next full
minute to inform the renderer that it's time to redraw the screen.
This way we no longer have any out of date times showing.