Show sunrise/-set on timeline
Renders the time before sunrise and after sunset as black (we're
pretty much assuming light theme right now).
Doesn't fit in with the pastel theme too well, but I think we'll
just put it behind an option later and then if a more minimalist
view is desired, that's how you get it.