Implement switching the current day with `h`/`l`
certainly WIP

Firstly we now have a `Day` type that is `{year,month,day}`.

Instead of the TUI Model maintaining a single "current" model, it
now maintains a map of already loaded models by day.
Similarly we have the awfully named `Program` which maintains a map
of file handler per day.

When we start the program we still load the current day, add the
FH and model to their respective maps.
Now when I press `h` though, it goes to the previous day.
The first time I go to this day it gets a new file handler for it,
loads the model and adds both the the maps.
If we go back to the first day with `l`, we already have that model
loaded, so no load from file necessary.

Tested loading and writing briefly, worked fine.