Show weather not just for current day
Firstly, this was wrong anyways, because it was using the actual,
real-life (accordint to `time.Now()`) current day even if a day was
specified with `-d`.

Now we keep not just that day's weather data, but all we get,
indexed now by (day, timestamp) instead of just timestamp.

Feel a bit dirty about just lazily creating the composite type
`DayAndTime` to allow for this indexing;
probably quite some room for restructuring that kind of code (day,
timestamp, ...) and naming these types more appropriately.