Draft week view
This really shows the necessity to redo the basic data model for the
TUI (`TUIModel`): It was conceived for the daily view only, thus
only kept one model for a day and when computing positions, it only
kept that one map of positions. This map is also indexed by event
IDs, which are unique for a single model (day) but not across
multiple days. I think we'll be able to keep something like
`ComputeRects()` (generified a bit more), but we'll have to redo the
TUI data model to support multiple views more cleanly.

ISSUES:
  - pretty sure hover state still exists from the day view and if
    you were to click in a certain place it would just do the edit
    on whatever event _would be_ under the cursor in that spot in
    the daily view, even if you're looking at another view
  - will crash unless all days of the week are intentionally loaded
    first by scrolling back and forward through all days of the week
    that would be shown; meaning: completely unusable
  - visually poor: names can easily overflow the boxes, ...
    probably don't show any detail beyond colored boxes in this view
    and then show details somewhere when events are hovered?