Implement `summarize` subcommand
Alternatively to this I was thinking about having multiple commands
as separate binaries in `cmd`. I tried this and it worked alright,
but as `go-flags` provides this subcommand mechanism I think this is
cleaner and less effort for further commands (although I had some
trouble with it, lacking documentation to understand it fully).