Wrap input processors when always-capture is desired
Instead of assuming in the modal overlay type that always-capture
was always desired, we now defer that decision to the applying
processor (either the base or topmost overlay).

Since the behavior is still desired, a simple wrapper that always
claims to capture is made available over any simple input processor.

I also don't know why we constructed those modal input processor
types in the controller which this commit replaces, it doesn't seem
to make sense, so its removed.