refactor: Restructure source code directories
I originally structured it this way because it was familiar from
other languages, made sense to me, and was the same structure that
FZF was using (I believe... or maybe it was gotop?).

I might want to create multiple binaries so the shift to 'cmd' makes
sense.

I also prefer the idiom of 'internal' meaning 'not for other use'.
I might move e.g. 'input' out of 'internal', there's bound to some
good code here somewhere (right...?), but for now its probably best
this way.