fix(files_provider): Avoid passing stored event pointers and fix event-date-inference
Previously, raw pointers for actually stored events were given, now
we clone events (still relatively cheap) to more easily spot errors
with updating.

Also, date inference based on end was wrong, should be on start
since we consider 2024-07-03 00:00 to be on 07-03 for start but the
same for 07-02 for end, so start is generally better for inference.